refactor(ir): `Lowered`'s node table is one MultiArrayList, its flags one EnumSet
ARCHITECTURE §6 phase 6, step 4 of 5: layout, no behaviour.

- `nodes: MultiArrayList(Node{name, kind, disc, dir})` replaces the four
  parallel lists node_order/node_kind/node_disciplines/node_dir. They
  were already columns; now one `append` keeps them one length, which
  only `appendNode`'s discipline did before.
- `uses: EnumSet(Kernel)` replaces the six `uses_*` bools and
  `reject_iteration_place`, a `?Ssa.Place` codegen only compared with
  null. The place goes back on `Lower`, and `.reject_iteration` is set
  where it is made.
- `HeldVar.place` and `LimitSlot.place` leave the output rows for
  `Lower.held_places`/`limit_places`, parallel lists nothing downstream
  can see. `Lowered` no longer imports ssa.zig.
- `num_ports`: usize -> u16, the node row index width.

Widths kept, each justified in tables.zig: a node row stays u16 and not
u8 because lowering must hold >256 unknowns for codegen's emitTopology to
count and refuse; `disc` stays a name because a node can name a
discipline `disciplines` has no row for, and an id would need its own
name table to save <=15 bytes on a <=256-row table.

@sizeOf(Lowered) 672 -> 584, 38 -> 29 fields. `Lower` 56 -> 59 fields
(the three places), @sizeOf(Lower) 1992 -> 1960.

A patch: goldens byte-identical, FAIL/XFAIL names unchanged. Measure D
only; A-C do not move.

+++ lib/root.zig
@@ -496,7 +496,7 @@ test "lint: source → MIR, arena freed clean" {
     defer res.deinit();
 
     try std.testing.expectEqualStrings("res", res.mir.name);
-    try std.testing.expectEqual(@as(usize, 2), res.lowered.num_ports);
+    try std.testing.expectEqual(@as(u16, 2), res.lowered.num_ports);
     try std.testing.expect(res.verdict.ok());
     try std.testing.expectEqual(res.unitCount(), res.verdict.unit_modes.len);
 }
